--- src/components/BlockLink.js
@@ -1,9 +1,0 @@
-import styled from '@emotion/styled'
-
-const BlockLink = styled.a`
-  display: block;
-  text-decoration: none;
-  color: inherit;
-`
-
-export default BlockLink
