--- src/components/Box.js
@@ -1,16 +1,0 @@
-import styled from '@emotion/styled'
-import {
-  space,
-  width,
-  flex,
-  color
-} from 'styled-system'
-
-const Box = styled.div`
-  ${space}
-  ${width}
-  ${flex}
-  ${color}
-`
-
-export default Box
