--- src/components/FigmaLogo.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const FigmaLogo = (props) => (
-<svg width="1779" height="1779" viewBox="0 0 1779 1779" style={{ width:'100%', height:'auto' }}><g transform="translate(2664 -1472)"><clipPath id="a" clipRule="evenodd"><path d="M-2664 1472h1779v1779h-1779V1472z" fill="#FFF"/></clipPath><g clipPath="url(#a)"><use xlinkHref="#path0_fill" transform="translate(-1775 2184)" fill="#1ABCFE"/><use xlinkHref="#path1_fill" transform="translate(-2130 2539)" fill="#0ACF83"/><use xlinkHref="#path2_fill" transform="matrix(0 1 1 0 -1775 1829)" fill="#FF7262"/><use xlinkHref="#path3_fill" transform="matrix(1 0 0 -1 -2130 2184)" fill="#F24E1E"/><use xlinkHref="#path3_fill" transform="matrix(1 0 0 -1 -2130 2539)" fill="#A259FF"/></g></g><defs><path id="path0_fill" d="M0 177.498C0 79.468 79.469 0 177.498 0h.002C275.531 0 355 79.47 355 177.5c0 98.031-79.469 177.5-177.5 177.5h-.002C79.468 355 0 275.531 0 177.502v-.004z"/><path id="path1_fill" d="M0 177.5C0 79.47 79.47 0 177.5 0H355v177.5c0 98.031-79.469 177.5-177.5 177.5C79.47 355 0 275.531 0 177.5z"/><path id="path2_fill" d="M0 0h355v177.5c0 98.031-79.469 177.5-177.5 177.5C79.47 355 0 275.531 0 177.5V0z"/><path id="path3_fill" d="M0 177.5C0 79.47 79.47 0 177.5 0H355v355H177.5C79.47 355 0 275.531 0 177.5z"/></defs></svg>
-)
-
-export default FigmaLogo
