--- src/components/FigmaMap.js
@@ -1,8 +1,0 @@
-import React from 'react'
-
-const FigmaMap = (props) => (
-<svg width="1400" height="800" viewBox="0 0 1400 800" style={{ width:'100%', height:'auto' }}>
-<g data-name="Canvas"><g data-name="NYC-America" clipPath="url(#clip0)" fill="none"><path fill="#111" d="M0 0h1400v800H0z"/><g data-name="Group 18"><g data-name="Group 17.1"><use xlinkHref="#path0" fill="#AAA" transform="translate(333.096 117)" data-name="Vector"/><use xlinkHref="#path1" fill="#AAA" transform="translate(366.775 117)" data-name="Vector"/><use xlinkHref="#path2" fill="#AAA" transform="translate(400.48 117)" data-name="Vector"/><use xlinkHref="#path3" fill="#AAA" transform="translate(434.192 117)" data-name="Vector"/><use xlinkHref="#path4" fill="#AAA" transform="translate(738.401 117)" data-name="Vector"/><use xlinkHref="#path5" fill="#AAA" transform="translate(333.096 150.671)" data-name="Vector"/><use xlinkHref="#path6" fill="#AAA" transform="translate(366.775 150.671)" data-name="Vector"/><use xlinkHref="#path7" fill="#AAA" transform="translate(400.48 150.671)" data-name="Vector"/><use xlinkHref="#path8" fill="#AAA" transform="translate(434.192 150.671)" data-name="Vector"/><use xlinkHref="#path9" fill="#AAA" transform="translate(467.865 150.671)" data-name="Vector"/><use xlinkHref="#path10" fill="#AAA" transform="translate(501.58 150.671)" data-name="Vector"/><use xlinkHref="#path11" fill="#AAA" transform="translate(536.201 150.671)" data-name="Vector"/><use xlinkHref="#path12" fill="#AAA" transform="translate(569.913 150.671)" data-name="Vector"/><use xlinkHref="#path9" fill="#AAA" transform="translate(603.589 150.671)" data-name="Vector"/><use xlinkHref="#path13" fill="#AAA" transform="translate(637.301 150.671)" data-name="Vector"/><use xlinkHref="#path14" fill="#AAA" transform="translate(670.977 150.671)" data-name="Vector"/><use xlinkHref="#path11" fill="#AAA" transform="translate(704.688 150.671)" data-name="Vector"/><use xlinkHref="#path12" fill="#AAA" transform="translate(738.401 150.671)" data-name="Vector"/><use xlinkHref="#path9" fill="#AAA" transform="translate(772.076 150.671)" data-name="Vector"/><use xlinkHref="#path13" fill="#AAA" transform="translate(805.785 150.671)" data-name="Vector"/><use xlinkHref="#path15" fill="#AAA" transform="translate(1109.99 150.671)" data-name="Vector"/><use xlinkHref="#path16" fill="#AAA" transform="translate(1143.67 150.671)" data-name="Vector"/><use xlinkHref="#path17" fill="#AAA" transform="translate(265.712 184.343)" data-name="Vector"/><use xlinkHref="#path18" fill="#AAA" transform="translate(299.388 184.343)" data-name="Vector"/><use xlinkHref="#path19" fill="#AAA" transform="translate(333.096 184.343)" data-name="Vector"/><use xlinkHref="#path20" fill="#AAA" transform="translate(366.775 184.343)" data-name="Vector"/><use xlinkHref="#path21" fill="#AAA" transform="translate(400.48 184.343)" data-name="Vector"/><use xlinkHref="#path22" fill="#AAA" transform="translate(434.192 184.343)" data-name="Vector"/><use xlinkHref="#path23" fill="#AAA" transform="translate(467.865 184.343)" data-name="Vector"/><use xlinkHref="#path24" fill="#AAA" transform="translate(501.58 184.343)" data-name="Vector"/><use xlinkHref="#path25" fill="#AAA" transform="translate(536.201 184.343)" data-name="Vector"/><use xlinkHref="#path26" fill="#AAA" transform="translate(569.913 184.343)" data-name="Vector"/><use xlinkHref="#path23" fill="#AAA" transform="translate(603.589 184.343)" data-name="Vector"/><use xlinkHref="#path27" fill="#AAA" transform="translate(637.301 184.343)" data-name="Vector"/><use xlinkHref="#path28" fill="#AAA" transform="translate(670.977 184.343)" data-name="Vector"/><use xlinkHref="#path25" fill="#AAA" transform="translate(704.688 184.343)" data-name="Vector"/><use xlinkHref="#path26" fill="#AAA" transform="translate(738.401 184.343)" data-name="Vector"/><use xlinkHref="#path23" fill="#AAA" transform="translate(772.076 184.343)" data-name="Vector"/><use xlinkHref="#path27" fill="#AAA" transform="translate(805.785 184.343)" data-name="Vector"/><use xlinkHref="#path29" fill="#AAA" transform="translate(839.457 184.343)" data-name="Vector"/><use xlinkHref="#path30" fill="#AAA" transform="translate(1042.6 184.343)" data-name="Vector"/><use xlinkHref="#path31" fill="#AAA" transform="translate(1076.28 184.343)" data-name="Vector"/><use xlinkHref="#path32" fill="#AAA" transform="translate(1109.99 184.343)" data-name="Vector"/><use xlinkHref="#path33" fill="#AAA" transform="translate(265.712 218.963)" data-name="Vector"/><use xlinkHref="#path34" fill="#AAA" transform="translate(299.388 218.963)" data-name="Vector"/><use xlinkHref="#path35" fill="#AAA" transform="translate(333.096 218.963)" data-name="Vector"/><use xlinkHref="#path36" fill="#AAA" transform="translate(366.775 218.963)" data-name="Vector"/><use xlinkHref="#path37" fill="#AAA" transform="translate(400.48 218.963)" data-name="Vector"/><use xlinkHref="#path38" fill="#AAA" transform="translate(434.192 218.963)" data-name="Vector"/><use xlinkHref="#path39" fill="#AAA" transform="translate(467.865 218.963)" data-name="Vector"/><use xlinkHref="#path40" fill="#AAA" transform="translate(501.58 218.963)" data-name="Vector"/><use xlinkHref="#path41" fill="#AAA" transform="translate(536.201 218.963)" data-name="Vector"/><use xlinkHref="#path42" fill="#AAA" transform="translate(569.913 218.963)" data-name="Vector"/><use xlinkHref="#path39" fill="#AAA" transform="translate(603.589 218.963)" data-name="Vector"/><use xlinkHref="#path43" fill="#AAA" transform="translate(637.301 218.963)" data-name="Vector"/><use xlinkHref="#path44" fill="#AAA" transform="translate(670.977 218.963)" data-name="Vector"/><use xlinkHref="#path41" fill="#AAA" transform="translate(704.688 218.963)" data-name="Vector"/><use xlinkHref="#path42" fill="#AAA" transform="translate(738.401 218.963)" data-name="Vector"/><use xlinkHref="#path39" fill="#AAA" transform="translate(772.076 218.963)" data-name="Vector"/><use xlinkHref="#path43" fill="#AAA" transform="translate(805.785 218.963)" data-name="Vector"/><use xlinkHref="#path45" fill="#AAA" transform="translate(839.457 218.963)" data-name="Vector"/><use xlinkHref="#path46" fill="#AAA" transform="translate(907.797 218.963)" data-name="Vector"/><use xlinkHref="#path47" fill="#AAA" transform="translate(1042.6 218.963)" data-name="Vector"/><use xlinkHref="#path46" fill="#AAA" transform="translate(1076.28 218.963)" data-name="Vector"/><use xlinkHref="#path48" fill="#AAA" transform="translate(1109.99 218.963)" data-name="Vector"/><use xlinkHref="#path49" fill="#AAA" transform="translate(232 252.636)" data-name="Vector"/><use xlinkHref="#path50" fill="#AAA" transform="translate(265.712 252.636)" data-name="Vector"/><use xlinkHref="#path51" fill="#AAA" transform="translate(299.388 252.636)" data-name="Vector"/><use xlinkHref="#path52" fill="#AAA" transform="translate(333.096 252.636)" data-name="Vector"/><use xlinkHref="#path53" fill="#AAA" transform="translate(366.775 252.636)" data-name="Vector"/><use xlinkHref="#path54" fill="#AAA" transform="translate(400.48 252.636)" data-name="Vector"/><use xlinkHref="#path55" fill="#AAA" transform="translate(434.192 252.636)" data-name="Vector"/><use xlinkHref="#path56" fill="#AAA" transform="translate(467.865 252.636)" data-name="Vector"/><use xlinkHref="#path57" fill="#AAA" transform="translate(501.58 252.636)" data-name="Vector"/><use xlinkHref="#path58" fill="#AAA" transform="translate(536.201 252.636)" data-name="Vector"/><use xlinkHref="#path59" fill="#AAA" transform="translate(569.913 252.636)" data-name="Vector"/><use xlinkHref="#path56" fill="#AAA" transform="translate(603.589 252.636)" data-name="Vector"/><use xlinkHref="#path60" fill="#AAA" transform="translate(637.301 252.636)" data-name="Vector"/><use xlinkHref="#path61" fill="#AAA" transform="translate(670.977 252.636)" data-name="Vector"/><use xlinkHref="#path58" fill="#AAA" transform="translate(704.688 252.636)" data-name="Vector"/><use xlinkHref="#path59" fill="#AAA" transform="translate(738.401 252.636)" data-name="Vector"/><use xlinkHref="#path56" fill="#AAA" transform="translate(772.076 252.636)" data-name="Vector"/><use xlinkHref="#path60" fill="#AAA" transform="translate(805.785 252.636)" data-name="Vector"/><use xlinkHref="#path62" fill="#AAA" transform="translate(839.457 252.636)" data-name="Vector"/><use xlinkHref="#path63" fill="#AAA" transform="translate(874.121 252.636)" data-name="Vector"/><use xlinkHref="#path64" fill="#AAA" transform="translate(907.797 252.636)" data-name="Vector"/><use xlinkHref="#path65" fill="#AAA" transform="translate(941.506 252.636)" data-name="Vector"/><use xlinkHref="#path66" fill="#AAA" transform="translate(1008.89 252.636)" data-name="Vector"/><use xlinkHref="#path63" fill="#AAA" transform="translate(1042.6 252.636)" data-name="Vector"/><use xlinkHref="#path64" fill="#DD0AFF" transform="translate(1076.28 252.636)" data-name="Vector"/><use xlinkHref="#path65" fill="#AAA" transform="translate(1109.99 252.636)" data-name="Vector"/><use xlinkHref="#path67" fill="#AAA" transform="translate(232 286.306)" data-name="Vector"/><use xlinkHref="#path68" fill="#AAA" transform="translate(265.712 286.306)" data-name="Vector"/><use xlinkHref="#path69" fill="#AAA" transform="translate(299.388 286.306)" data-name="Vector"/><use xlinkHref="#path70" fill="#AAA" transform="translate(333.096 286.306)" data-name="Vector"/><use xlinkHref="#path71" fill="#AAA" transform="translate(366.775 286.306)" data-name="Vector"/><use xlinkHref="#path72" fill="#AAA" transform="translate(400.48 286.306)" data-name="Vector"/><use xlinkHref="#path73" fill="#AAA" transform="translate(434.192 286.306)" data-name="Vector"/><use xlinkHref="#path74" fill="#AAA" transform="translate(467.865 286.306)" data-name="Vector"/><use xlinkHref="#path75" fill="#AAA" transform="translate(501.58 286.306)" data-name="Vector"/><use xlinkHref="#path76" fill="#AAA" transform="translate(536.201 286.306)" data-name="Vector"/><use xlinkHref="#path77" fill="#AAA" transform="translate(569.913 286.306)" data-name="Vector"/><use xlinkHref="#path74" fill="#AAA" transform="translate(603.589 286.306)" data-name="Vector"/><use xlinkHref="#path78" fill="#AAA" transform="translate(637.301 286.306)" data-name="Vector"/><use xlinkHref="#path79" fill="#AAA" transform="translate(670.977 286.306)" data-name="Vector"/><use xlinkHref="#path76" fill="#AAA" transform="translate(704.688 286.306)" data-name="Vector"/><use xlinkHref="#path77" fill="#AAA" transform="translate(738.401 286.306)" data-name="Vector"/><use xlinkHref="#path74" fill="#AAA" transform="translate(772.076 286.306)" data-name="Vector"/><use xlinkHref="#path78" fill="#AAA" transform="translate(805.785 286.306)" data-name="Vector"/><use xlinkHref="#path80" fill="#AAA" transform="translate(839.457 286.306)" data-name="Vector"/><use xlinkHref="#path81" fill="#AAA" transform="translate(874.121 286.306)" data-name="Vector"/><use xlinkHref="#path82" fill="#AAA" transform="translate(907.797 286.306)" data-name="Vector"/><use xlinkHref="#path83" fill="#AAA" transform="translate(941.506 286.306)" data-name="Vector"/><use xlinkHref="#path80" fill="#AAA" transform="translate(975.185 286.306)" data-name="Vector"/><use xlinkHref="#path84" fill="#AAA" transform="translate(1008.89 286.306)" data-name="Vector"/><use xlinkHref="#path81" fill="#AAA" transform="translate(1042.6 286.306)" data-name="Vector"/><use xlinkHref="#path82" fill="#AAA" transform="translate(1076.28 286.306)" data-name="Vector"/><use xlinkHref="#path85" fill="#AAA" transform="translate(232 319.979)" data-name="Vector"/><use xlinkHref="#path86" fill="#AAA" transform="translate(265.712 319.979)" data-name="Vector"/><use xlinkHref="#path87" fill="#AAA" transform="translate(299.388 319.979)" data-name="Vector"/><use xlinkHref="#path88" fill="#AAA" transform="translate(333.096 319.979)" data-name="Vector"/><use xlinkHref="#path89" fill="#AAA" transform="translate(366.775 319.979)" data-name="Vector"/><use xlinkHref="#path90" fill="#AAA" transform="translate(400.48 319.979)" data-name="Vector"/><use xlinkHref="#path91" fill="#AAA" transform="translate(434.192 319.979)" data-name="Vector"/><use xlinkHref="#path92" fill="#AAA" transform="translate(467.865 319.979)" data-name="Vector"/><use xlinkHref="#path93" fill="#AAA" transform="translate(501.58 319.979)" data-name="Vector"/><use xlinkHref="#path94" fill="#AAA" transform="translate(536.201 319.979)" data-name="Vector"/><use xlinkHref="#path95" fill="#AAA" transform="translate(569.913 319.979)" data-name="Vector"/><use xlinkHref="#path92" fill="#AAA" transform="translate(603.589 319.979)" data-name="Vector"/><use xlinkHref="#path96" fill="#AAA" transform="translate(637.301 319.979)" data-name="Vector"/><use xlinkHref="#path97" fill="#AAA" transform="translate(670.977 319.979)" data-name="Vector"/><use xlinkHref="#path94" fill="#AAA" transform="translate(704.688 319.979)" data-name="Vector"/><use xlinkHref="#path95" fill="#AAA" transform="translate(738.401 319.979)" data-name="Vector"/><use xlinkHref="#path92" fill="#AAA" transform="translate(772.076 319.979)" data-name="Vector"/><use xlinkHref="#path96" fill="#AAA" transform="translate(805.785 319.979)" data-name="Vector"/><use xlinkHref="#path98" fill="#AAA" transform="translate(839.457 319.979)" data-name="Vector"/><use xlinkHref="#path99" fill="#AAA" transform="translate(874.121 319.979)" data-name="Vector"/><use xlinkHref="#path100" fill="#AAA" transform="translate(907.797 319.979)" data-name="Vector"/><use xlinkHref="#path101" fill="#AAA" transform="translate(941.506 319.979)" data-name="Vector"/><use xlinkHref="#path98" fill="#AAA" transform="translate(975.185 319.979)" data-name="Vector"/><use xlinkHref="#path102" fill="#AAA" transform="translate(1008.89 319.979)" data-name="Vector"/><use xlinkHref="#path99" fill="#AAA" transform="translate(1042.6 319.979)" data-name="Vector"/><use xlinkHref="#path100" fill="#AAA" transform="translate(1076.28 319.979)" data-name="Vector"/><use xlinkHref="#path103" fill="#AAA" transform="translate(265.712 353.667)" data-name="Vector"/><use xlinkHref="#path104" fill="#AAA" transform="translate(299.388 353.667)" data-name="Vector"/><use xlinkHref="#path105" fill="#AAA" transform="translate(333.096 353.667)" data-name="Vector"/><use xlinkHref="#path106" fill="#AAA" transform="translate(366.775 353.667)" data-name="Vector"/><use xlinkHref="#path107" fill="#AAA" transform="translate(400.48 353.667)" data-name="Vector"/><use xlinkHref="#path108" fill="#AAA" transform="translate(434.192 353.667)" data-name="Vector"/><use xlinkHref="#path109" fill="#AAA" transform="translate(467.865 353.667)" data-name="Vector"/><use xlinkHref="#path110" fill="#AAA" transform="translate(501.58 353.667)" data-name="Vector"/><use xlinkHref="#path111" fill="#AAA" transform="translate(536.201 353.667)" data-name="Vector"/><use xlinkHref="#path112" fill="#AAA" transform="translate(569.913 353.667)" data-name="Vector"/><use xlinkHref="#path109" fill="#AAA" transform="translate(603.589 353.667)" data-name="Vector"/><use xlinkHref="#path113" fill="#AAA" transform="translate(637.301 353.667)" data-name="Vector"/><use xlinkHref="#path114" fill="#AAA" transform="translate(670.977 353.667)" data-name="Vector"/><use xlinkHref="#path111" fill="#AAA" transform="translate(704.688 353.667)" data-name="Vector"/><use xlinkHref="#path112" fill="#AAA" transform="translate(738.401 353.667)" data-name="Vector"/><use xlinkHref="#path109" fill="#AAA" transform="translate(772.076 353.667)" data-name="Vector"/><use xlinkHref="#path113" fill="#AAA" transform="translate(805.785 353.667)" data-name="Vector"/><use xlinkHref="#path115" fill="#AAA" transform="translate(839.457 353.667)" data-name="Vector"/><use xlinkHref="#path116" fill="#AAA" transform="translate(874.121 353.667)" data-name="Vector"/><use xlinkHref="#path117" fill="#AAA" transform="translate(907.797 353.667)" data-name="Vector"/><use xlinkHref="#path118" fill="#AAA" transform="translate(941.506 353.667)" data-name="Vector"/><use xlinkHref="#path115" fill="#AAA" transform="translate(975.185 353.667)" data-name="Vector"/><use xlinkHref="#path119" fill="#AAA" transform="translate(1008.89 353.667)" data-name="Vector"/><use xlinkHref="#path116" fill="#AAA" transform="translate(1042.6 353.667)" data-name="Vector"/><use xlinkHref="#path117" fill="#AAA" transform="translate(1076.28 353.667)" data-name="Vector"/><use xlinkHref="#path33" fill="#AAA" transform="translate(265.712 387.338)" data-name="Vector"/><use xlinkHref="#path34" fill="#AAA" transform="translate(299.388 387.338)" data-name="Vector"/><use xlinkHref="#path35" fill="#AAA" transform="translate(333.096 387.338)" data-name="Vector"/><use xlinkHref="#path36" fill="#AAA" transform="translate(366.775 387.338)" data-name="Vector"/><use xlinkHref="#path37" fill="#AAA" transform="translate(400.48 387.338)" data-name="Vector"/><use xlinkHref="#path38" fill="#AAA" transform="translate(434.192 387.338)" data-name="Vector"/><use xlinkHref="#path39" fill="#AAA" transform="translate(467.865 387.338)" data-name="Vector"/><use xlinkHref="#path40" fill="#AAA" transform="translate(501.58 387.338)" data-name="Vector"/><use xlinkHref="#path41" fill="#AAA" transform="translate(536.201 387.338)" data-name="Vector"/><use xlinkHref="#path42" fill="#AAA" transform="translate(569.913 387.338)" data-name="Vector"/><use xlinkHref="#path39" fill="#AAA" transform="translate(603.589 387.338)" data-name="Vector"/><use xlinkHref="#path43" fill="#AAA" transform="translate(637.301 387.338)" data-name="Vector"/><use xlinkHref="#path44" fill="#AAA" transform="translate(670.977 387.338)" data-name="Vector"/><use xlinkHref="#path41" fill="#AAA" transform="translate(704.688 387.338)" data-name="Vector"/><use xlinkHref="#path42" fill="#AAA" transform="translate(738.401 387.338)" data-name="Vector"/><use xlinkHref="#path39" fill="#AAA" transform="translate(772.076 387.338)" data-name="Vector"/><use xlinkHref="#path43" fill="#AAA" transform="translate(805.785 387.338)" data-name="Vector"/><use xlinkHref="#path45" fill="#AAA" transform="translate(839.457 387.338)" data-name="Vector"/><use xlinkHref="#path47" fill="#AAA" transform="translate(874.121 387.338)" data-name="Vector"/><use xlinkHref="#path46" fill="#AAA" transform="translate(907.797 387.338)" data-name="Vector"/><use xlinkHref="#path48" fill="#AAA" transform="translate(941.506 387.338)" data-name="Vector"/><use xlinkHref="#path45" fill="#AAA" transform="translate(975.185 387.338)" data-name="Vector"/><use xlinkHref="#path120" fill="#AAA" transform="translate(1008.89 387.338)" data-name="Vector"/><use xlinkHref="#path47" fill="#AAA" transform="translate(1042.6 387.338)" data-name="Vector"/><use xlinkHref="#path46" fill="#AAA" transform="translate(1076.28 387.338)" data-name="Vector"/><use xlinkHref="#path51" fill="#AAA" transform="translate(299.388 421.009)" data-name="Vector"/><use xlinkHref="#path52" fill="#AAA" transform="translate(333.096 421.009)" data-name="Vector"/><use xlinkHref="#path53" fill="#AAA" transform="translate(366.775 421.009)" data-name="Vector"/><use xlinkHref="#path54" fill="#AAA" transform="translate(400.48 421.009)" data-name="Vector"/><use xlinkHref="#path55" fill="#AAA" transform="translate(434.192 421.009)" data-name="Vector"/><use xlinkHref="#path56" fill="#AAA" transform="translate(467.865 421.009)" data-name="Vector"/><use xlinkHref="#path57" fill="#AAA" transform="translate(501.58 421.009)" data-name="Vector"/><use xlinkHref="#path58" fill="#AAA" transform="translate(536.201 421.009)" data-name="Vector"/><use xlinkHref="#path59" fill="#AAA" transform="translate(569.913 421.009)" data-name="Vector"/><use xlinkHref="#path56" fill="#AAA" transform="translate(603.589 421.009)" data-name="Vector"/><use xlinkHref="#path60" fill="#AAA" transform="translate(637.301 421.009)" data-name="Vector"/><use xlinkHref="#path61" fill="#AAA" transform="translate(670.977 421.009)" data-name="Vector"/><use xlinkHref="#path58" fill="#AAA" transform="translate(704.688 421.009)" data-name="Vector"/><use xlinkHref="#path59" fill="#AAA" transform="translate(738.401 421.009)" data-name="Vector"/><use xlinkHref="#path56" fill="#AAA" transform="translate(772.076 421.009)" data-name="Vector"/><use xlinkHref="#path60" fill="#AAA" transform="translate(805.785 421.009)" data-name="Vector"/><use xlinkHref="#path62" fill="#AAA" transform="translate(839.457 421.009)" data-name="Vector"/><use xlinkHref="#path63" fill="#AAA" transform="translate(874.121 421.009)" data-name="Vector"/><use xlinkHref="#path64" fill="#AAA" transform="translate(907.797 421.009)" data-name="Vector"/><use xlinkHref="#path65" fill="#AAA" transform="translate(941.506 421.009)" data-name="Vector"/><use xlinkHref="#path62" fill="#AAA" transform="translate(975.185 421.009)" data-name="Vector"/><use xlinkHref="#path66" fill="#AAA" transform="translate(1008.89 421.009)" data-name="Vector"/><use xlinkHref="#path63" fill="#AAA" transform="translate(1042.6 421.009)" data-name="Vector"/><use xlinkHref="#path64" fill="#AAA" transform="translate(1076.28 421.009)" data-name="Vector"/><use xlinkHref="#path69" fill="#AAA" transform="translate(299.388 454.683)" data-name="Vector"/><use xlinkHref="#path70" fill="#AAA" transform="translate(333.096 454.683)" data-name="Vector"/><use xlinkHref="#path71" fill="#AAA" transform="translate(366.775 454.683)" data-name="Vector"/><use xlinkHref="#path72" fill="#AAA" transform="translate(400.48 454.683)" data-name="Vector"/><use xlinkHref="#path73" fill="#AAA" transform="translate(434.192 454.683)" data-name="Vector"/><use xlinkHref="#path74" fill="#AAA" transform="translate(467.865 454.683)" data-name="Vector"/><use xlinkHref="#path75" fill="#AAA" transform="translate(501.58 454.683)" data-name="Vector"/><use xlinkHref="#path76" fill="#AAA" transform="translate(536.201 454.683)" data-name="Vector"/><use xlinkHref="#path77" fill="#AAA" transform="translate(569.913 454.683)" data-name="Vector"/><use xlinkHref="#path74" fill="#AAA" transform="translate(603.589 454.683)" data-name="Vector"/><use xlinkHref="#path78" fill="#AAA" transform="translate(637.301 454.683)" data-name="Vector"/><use xlinkHref="#path79" fill="#AAA" transform="translate(670.977 454.683)" data-name="Vector"/><use xlinkHref="#path76" fill="#AAA" transform="translate(704.688 454.683)" data-name="Vector"/><use xlinkHref="#path77" fill="#AAA" transform="translate(738.401 454.683)" data-name="Vector"/><use xlinkHref="#path74" fill="#AAA" transform="translate(772.076 454.683)" data-name="Vector"/><use xlinkHref="#path78" fill="#AAA" transform="translate(805.785 454.683)" data-name="Vector"/><use xlinkHref="#path80" fill="#AAA" transform="translate(839.457 454.683)" data-name="Vector"/><use xlinkHref="#path81" fill="#AAA" transform="translate(874.121 454.683)" data-name="Vector"/><use xlinkHref="#path82" fill="#AAA" transform="translate(907.797 454.683)" data-name="Vector"/><use xlinkHref="#path83" fill="#AAA" transform="translate(941.506 454.683)" data-name="Vector"/><use xlinkHref="#path80" fill="#AAA" transform="translate(975.185 454.683)" data-name="Vector"/><use xlinkHref="#path84" fill="#AAA" transform="translate(1008.89 454.683)" data-name="Vector"/><use xlinkHref="#path81" fill="#AAA" transform="translate(1042.6 454.683)" data-name="Vector"/><use xlinkHref="#path88" fill="#AAA" transform="translate(333.096 488.354)" data-name="Vector"/><use xlinkHref="#path89" fill="#AAA" transform="translate(366.775 488.354)" data-name="Vector"/><use xlinkHref="#path90" fill="#AAA" transform="translate(400.48 488.354)" data-name="Vector"/><use xlinkHref="#path91" fill="#AAA" transform="translate(434.192 488.354)" data-name="Vector"/><use xlinkHref="#path92" fill="#AAA" transform="translate(467.865 488.354)" data-name="Vector"/><use xlinkHref="#path93" fill="#AAA" transform="translate(501.58 488.354)" data-name="Vector"/><use xlinkHref="#path94" fill="#AAA" transform="translate(536.201 488.354)" data-name="Vector"/><use xlinkHref="#path95" fill="#AAA" transform="translate(569.913 488.354)" data-name="Vector"/><use xlinkHref="#path92" fill="#AAA" transform="translate(603.589 488.354)" data-name="Vector"/><use xlinkHref="#path96" fill="#AAA" transform="translate(637.301 488.354)" data-name="Vector"/><use xlinkHref="#path97" fill="#AAA" transform="translate(670.977 488.354)" data-name="Vector"/><use xlinkHref="#path94" fill="#AAA" transform="translate(704.688 488.354)" data-name="Vector"/><use xlinkHref="#path95" fill="#AAA" transform="translate(738.401 488.354)" data-name="Vector"/><use xlinkHref="#path92" fill="#AAA" transform="translate(772.076 488.354)" data-name="Vector"/><use xlinkHref="#path96" fill="#AAA" transform="translate(805.785 488.354)" data-name="Vector"/><use xlinkHref="#path98" fill="#AAA" transform="translate(839.457 488.354)" data-name="Vector"/><use xlinkHref="#path99" fill="#AAA" transform="translate(874.121 488.354)" data-name="Vector"/><use xlinkHref="#path100" fill="#AAA" transform="translate(907.797 488.354)" data-name="Vector"/><use xlinkHref="#path101" fill="#AAA" transform="translate(941.506 488.354)" data-name="Vector"/><use xlinkHref="#path98" fill="#AAA" transform="translate(975.185 488.354)" data-name="Vector"/><use xlinkHref="#path102" fill="#AAA" transform="translate(1008.89 488.354)" data-name="Vector"/><use xlinkHref="#path121" fill="#AAA" transform="translate(434.192 522.043)" data-name="Vector"/><use xlinkHref="#path121" fill="#AAA" transform="translate(401 522.043)" data-name="Vector"/><use xlinkHref="#path122" fill="#AAA" transform="translate(467.865 522.043)" data-name="Vector"/><use xlinkHref="#path123" fill="#AAA" transform="translate(501.58 522.043)" data-name="Vector"/><use xlinkHref="#path124" fill="#AAA" transform="translate(536.201 522.043)" data-name="Vector"/><use xlinkHref="#path125" fill="#AAA" transform="translate(569.913 522.043)" data-name="Vector"/><use xlinkHref="#path122" fill="#AAA" transform="translate(603.589 522.043)" data-name="Vector"/><use xlinkHref="#path126" fill="#AAA" transform="translate(637.301 522.043)" data-name="Vector"/><use xlinkHref="#path127" fill="#AAA" transform="translate(670.977 522.043)" data-name="Vector"/><use xlinkHref="#path124" fill="#AAA" transform="translate(704.688 522.043)" data-name="Vector"/><use xlinkHref="#path125" fill="#AAA" transform="translate(738.401 522.043)" data-name="Vector"/><use xlinkHref="#path122" fill="#AAA" transform="translate(772.076 522.043)" data-name="Vector"/><use xlinkHref="#path126" fill="#AAA" transform="translate(805.785 522.043)" data-name="Vector"/><use xlinkHref="#path128" fill="#AAA" transform="translate(839.457 522.043)" data-name="Vector"/><use xlinkHref="#path129" fill="#AAA" transform="translate(874.121 522.043)" data-name="Vector"/><use xlinkHref="#path130" fill="#AAA" transform="translate(907.797 522.043)" data-name="Vector"/><use xlinkHref="#path131" fill="#AAA" transform="translate(941.506 522.043)" data-name="Vector"/><use xlinkHref="#path128" fill="#AAA" transform="translate(975.185 522.043)" data-name="Vector"/><use xlinkHref="#path132" fill="#AAA" transform="translate(1008.89 522.043)" data-name="Vector"/><use xlinkHref="#path133" fill="#AAA" transform="translate(569.913 555.715)" data-name="Vector"/><use xlinkHref="#path134" fill="#AAA" transform="translate(603.589 555.715)" data-name="Vector"/><use xlinkHref="#path135" fill="#AAA" transform="translate(637.301 555.715)" data-name="Vector"/><use xlinkHref="#path136" fill="#AAA" transform="translate(670.977 555.715)" data-name="Vector"/><use xlinkHref="#path137" fill="#AAA" transform="translate(704.688 555.715)" data-name="Vector"/><use xlinkHref="#path133" fill="#AAA" transform="translate(738.401 555.715)" data-name="Vector"/><use xlinkHref="#path134" fill="#AAA" transform="translate(772.076 555.715)" data-name="Vector"/><use xlinkHref="#path135" fill="#AAA" transform="translate(805.785 555.715)" data-name="Vector"/><use xlinkHref="#path138" fill="#AAA" transform="translate(839.457 555.715)" data-name="Vector"/><use xlinkHref="#path139" fill="#AAA" transform="translate(874.121 555.715)" data-name="Vector"/><use xlinkHref="#path140" fill="#AAA" transform="translate(941.506 555.715)" data-name="Vector"/><use xlinkHref="#path138" fill="#AAA" transform="translate(975.185 555.715)" data-name="Vector"/><use xlinkHref="#path141" fill="#AAA" transform="translate(1008.89 555.715)" data-name="Vector"/><use xlinkHref="#path142" fill="#AAA" transform="translate(637.301 590.316)" data-name="Vector"/><use xlinkHref="#path143" fill="#AAA" transform="translate(670.977 590.316)" data-name="Vector"/><use xlinkHref="#path144" fill="#AAA" transform="translate(704.688 590.316)" data-name="Vector"/><use xlinkHref="#path145" fill="#AAA" transform="translate(738.401 590.316)" data-name="Vector"/><use xlinkHref="#path146" fill="#AAA" transform="translate(1008.89 590.316)" data-name="Vector"/><use xlinkHref="#path147" fill="#AAA" transform="translate(1042.6 590.316)" data-name="Vector"/><use xlinkHref="#path148" fill="#AAA" transform="translate(670.977 623.99)" data-name="Vector"/><use xlinkHref="#path149" fill="#AAA" transform="translate(704.688 623.99)" data-name="Vector"/><use xlinkHref="#path150" fill="#AAA" transform="translate(1042.6 623.99)" data-name="Vector"/><use xlinkHref="#path124" fill="#AAA" transform="translate(704.688 657.679)" data-name="Vector"/></g><g data-name="Group 15.1"><g data-name="move-tool"><use xlinkHref="#path151" fill="#000" transform="translate(1122 288.999)"/><use xlinkHref="#path152" stroke="#fff" strokeWidth="5" strokeLinecap="square" transform="translate(1122 288.999)"/></g><g data-name="Group 11.1"><path fill="#00F" d="M1178 375.68h80v-44h-80z" data-name="Rectangle 19"/></g></g></g></g></g><defs><clipPath id="clip0"><path fill="#fff" d="M0 0h1400v800H0z"/></clipPath><path id="path0" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path1" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path2" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path3" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path4" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path5" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path6" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path7" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path8" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path9" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path10" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path11" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path12" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path13" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path14" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path15" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path16" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path17" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path18" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path19" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path20" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path21" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path22" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path23" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path24" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path25" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path26" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path27" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path28" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path29" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path30" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path31" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path32" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path33" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path34" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path35" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path36" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path37" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path38" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path39" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path40" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path41" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path42" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path43" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path44" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path45" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path46" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path47" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path48" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path49" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path50" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path51" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path52" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path53" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path54" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path55" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path56" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path57" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path58" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path59" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path60" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path61" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path62" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path63" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path64" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path65" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path66" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path67" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path68" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path69" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path70" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path71" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path72" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path73" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path74" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path75" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path76" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path77" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path78" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path79" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path80" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path81" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path82" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path83" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path84" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path85" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path86" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path87" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path88" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path89" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path90" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path91" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path92" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path93" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path94" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path95" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path96" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path97" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path98" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path99" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path100" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path101" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path102" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path103" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path104" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path105" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path106" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path107" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path108" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path109" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path110" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path111" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path112" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path113" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path114" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path115" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path116" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path117" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path118" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path119" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path120" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path121" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path122" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path123" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path124" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path125" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path126" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path127" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path128" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path129" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path130" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path131" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path132" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path133" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path134" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path135" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path136" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path137" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path138" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path139" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path140" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path141" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path142" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path143" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path144" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path145" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path146" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path147" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path148" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path149" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path150" d="M12.168 24.32c6.72 0 12.168-5.444 12.168-12.16S18.89 0 12.168 0C5.448 0 0 5.444 0 12.16s5.448 12.16 12.168 12.16z"/><path id="path151" d="M11.472 55.68L0 0l48.005 29.067-23.096 6.279L11.472 55.68z"/><path id="path152" d="M0 0l11.472 55.68 13.437-20.334 23.096-6.279L0 0z"/></defs></svg>
-)
-
-export default FigmaMap
