--- src/components/Image.js
@@ -1,14 +1,0 @@
-import styled from '@emotion/styled'
-import {
-  space,
-  width
-} from 'styled-system'
-
-const Image = styled.img`
-  max-width: 100%;
-  height: auto;
-  ${space}
-  ${width}
-`
-
-export default Image
