--- src/components/PastEventCard.js
@@ -1,48 +1,0 @@
-import React from 'react'
-import Box from './Box'
-import Flex from './Flex'
-import BackgroundImage from './BackgroundImage'
-import Heading from './Heading'
-import Text from './Text'
-import Caps from './Caps'
-import Link from './Link'
-
-export default ({
-  image,
-  permalink,
-  title,
-  date,
-  location,
-  description,
-}) => (
-  <Flex
-    flexWrap='wrap'
-    mx={-3}
-    alignItems='center'>
-    <Box
-      px={3}
-      flex='none'
-      width={[ 1, 1/2, 2/5 ]}>
-      <BackgroundImage
-        src={image}
-        bg='gray'
-        ratio={3/5}
-      />
-    </Box>
-    <Box px={3} my={3} width={[ 1, 1/2, 3/5 ]}>
-      <Link href={permalink}>
-      <Heading
-        fontSize={[ 4, 5 ]}
-        mb={2}>
-        {title}
-      </Heading>
-      </Link>
-      <Caps mb={3}>
-        {date} @ {location}
-      </Caps>
-      <Text mb={3}>
-        {description}
-      </Text>
-    </Box>
-  </Flex>
-)
