--- src/components/Text.js
@@ -1,18 +1,0 @@
-import styled from '@emotion/styled'
-import {
-  space,
-  color,
-  fontSize,
-  fontWeight,
-  lineHeight,
-} from 'styled-system'
-
-const Text = styled.div`
-  ${space}
-  ${color}
-  ${fontSize}
-  ${fontWeight}
-  ${lineHeight}
-`
-
-export default Text
