+++ src/pages/footer.mdx
@@ -1,6 +1,5 @@
 import { ThemeProvider } from 'theme-ui'
 import Container from '../components/Container'
-import Box from '../components/Box'
 
 export default props =>
   <ThemeProvider
