+++ src/pages/header.mdx
@@ -1,31 +1,7 @@
-import { ThemeProvider } from 'theme-ui'
+import Header from '../blocks/Header'
 import ColorModeButton from '../components/ColorModeButton'
-import Container from '../components/Container'
 
-export default props =>
-  <ThemeProvider
-    theme={{
-      styles: {
-        a: {
-          color: 'inherit',
-          fontWeight: 'bold',
-          textDecoration: 'none',
-          ':hover': {
-            color: 'primary',
-          },
-        },
-      }
-    }}>
-    <Container
-      {...props}
-      css={{
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'space-between',
-        py: 3,
-      }}
-    />
-  </ThemeProvider>
+export default Header
 
 [DSC](/)
