+++ src/pages/hosting-guidelines.mdx
@@ -1,0 +1,42 @@
+import Layout from '../layout'
+
+export default Layout
+
+# Hosting guidelines
+
+Guidelines for hosting a Design Systems Coalition NYC meetup
+
+## About us
+The [Design Systems Coalition](https://www.meetup.com/NYC-Design-Systems-Coalition/) is a niche meetup for people enthusiastic about design systems. This includes professionals, freelancers, designers and developers, brand designers, managers and directors, product managers, and anyone else interested in learning more or sharing their own experiences with building design systems.
+
+The NYC meetup is organized by [Diana Mounter](https://twitter.com/broccolini), Design Systems Manager at [GitHub](https://github.com/), [Tom Takigayama](https://twitter.com/tomtaki), Design Lead at [Justworks]([https://justworks.com](https://justworks.com/)), and [Brent Jackson](https://twitter.com/jxnblk), Principle UX developer at [Priceline.com](https://www.priceline.com/). The meetup first formed in San Francisco and has chapters popping up in cities around the globe. It was born out of the [Design Systems slack](http://designsystems.herokuapp.com/) community started by [Jina Anne](https://twitter.com/jina). Jina is a design systems advocate, former design lead on the [Lightning Design System](https://www.lightningdesignsystem.com/), and organizer of [Clarity](http://clarityconf.com/).
+
+## Event format
+
+Each meetup features 3 speakers giving 10 minute lightning talks on the same topic area. After the talks have finished we kick off a panel-style Q&A session with curated questions, and then open questions to the floor. Our goal with the lightning talks and Q&A is to stimulate discussion, keeping plenty of time for networking and meeting other design systems enthusiasts. We focus each meetup on one topic area so that it's easier for speakers to create a short and focussed talk, and to ensure that over time, we cover the many areas that design systems work involves.
+
+## Venue logistics
+- The venue should be able to accommodate between 60—100 people. Free meetups usually attract a 50% attendance rate, therefore we offer more spots on Meetup than we expect to attend. We'll work with you to ensure the numbers of attendees works for your venue capacity.
+- AV equipement will be needed for projecting presenter slides.
+- Not required, but if you are able to record the talks so that they can be shared online after the event that would be appreciated. Please let us know before hand if you are able to do this so we can notify presenters and attendees that talks will be recorded.
+- A primary goal of Design Systems Coalition NYC is to be inclusive to the largest number of members, this means we require venues to be accessible, safe, and welcoming environments. Please read our [Code of Conduct](https://github.com/design-systems-coalition-NYC/meetup/blob/master/CODE_OF_CONDUCT.md) for more information and reach out if you have any concerns about meeting these requirements.
+
+## Food and drink
+- Hosts are expected to provide food and drinks for attendees since the event will be in the evening. Snacks should include vegan and gluten-free options, and if providing alcohol, non-alcoholic options should also be available. In the past hosts have provided foos such as charcuterrie style platters, chips and dips, cheeses & crackers, cookies, and pizza. Most hosts provide wine, beer, and a soda option, as well as water. We don't expect you to provide enough to replace an evening meal, just enough to tide folks over if they haven't eaten before they get to the venue.
+
+## Time and length
+- We'd like meetups to begin at 6.30, with talks beginning around 7, and the event ending around 9pm. We can be bit flexible on this though so please let us know if you have constraints around timing.
+- We'd like organizers to be able to have access to the venue about 30 minutes before hand to help set up and make sure everything runs smoothly.
+
+## Venue access and registration
+- Tickets: the event is not ticketed but we will have an rsvp list which we will check on the door. We won't turn un-registered people away unless space is limited or the venue has other access restrictions. Usually there will be a number of no-shows that balance out occasional extras that turn up. We will take the name and email of un-registered attendees.
+- Registration: if the venue requires a security check with ID before entering the event, please let us know so we can notify attendees. Organizers will need a small table near the entry to the event room for checking off RSVP's as people arrive.
+- Building access: a primary goal of our meetup is to be inclusive to the largest number of members, this means we require venues to be accessible, safe, and welcoming environments.
+
+## Host
+We would love someone from the hosting company to present on their design system, or a related topic. Each event will be focussed around a niche topic and we'll work with you to find a topic that works for you and is someone the design systems community is interested in.
+
+As part of hosting, we'd like to offer the following:
+- Logos on event materials (the meetup page, slides loop during event, youtube video if the event is recorded)
+- If you have jobs advertized that you'd like to share at the meetup we'll add the details to the slides loop, and you can give a shoutout at the event, we'll post on our [twitter](https://twitter.com/nyc_dsc) too if you like.
+- If there's something else you'd like to do that would make it worthwhile you hosting please let us know and we can discuss ideas.
