+++ src/pages/index.mdx
@@ -3,6 +3,7 @@ import Hero from './hero.mdx'
 import NextEvent from './next-event.mdx'
 import Speakers from './speakers.mdx'
 import CFP from './cfp.mdx'
+import PastEvents from './past-events.mdx'
 import HostPromo from './host-promo.mdx'
 import Organizers from './organizers.mdx'
 import Footer from './footer.mdx'
@@ -12,6 +13,7 @@ import Footer from './footer.mdx'
 <NextEvent />
 <Speakers />
 <CFP />
+<PastEvents />
 <HostPromo />
 <Organizers />
 <Footer />
