+++ src/pages/old-style-guide.mdx
@@ -1,0 +1,15 @@
+
+# Style Guide
+
+<iframe
+  title='figma'
+  style={{
+    borderWidth: '4px',
+    borderStyle: 'solid',
+    borderColor: '#eee',
+    width:'800px',
+    height:'450px'
+  }}
+  src='https://www.figma.com/embed?embed_host=share&url=https://www.figma.com/file/QaYbQ2LGdA3lefOWZCosDcrt/DSC-NYC-logos'
+  allowFullScreen
+/>
