+++ src/pages/past-events.mdx
@@ -1,120 +1,115 @@
+import Section from '../blocks/Section'
+import Cards from '../blocks/Cards'
+
+export default Section
 
 ## Past Events
 
-### Happy Hour!
+<Cards>
 
 ![](https://user-images.githubusercontent.com/334891/45930307-7c51c500-bf2c-11e8-97ce-0e1c2231c76f.jpg)
+
+### Happy Hour!
 
 March 27 @ The Winslow
 
 Join us for a happy hour and meet other design systems professionals and enthusiasts. This event will give folks the opportunity to meet up in a casual setting.
 [Meetup](https://www.meetup.com/NYC-Design-Systems-Coalition/events/259942025/)
 
----
-
-### Happy Hour!
 
 ![](https://user-images.githubusercontent.com/334891/41512111-a336b338-7251-11e8-93b8-7371c8e50a39.JPG)
+
+### Happy Hour!
 
 Jan 23 @ The Winslow
 
 Join us for a happy hour and meet other design systems professionals and enthusiasts. This event will give folks the opportunity to meet up in a casual setting.
 [Meetup](https://www.meetup.com/NYC-Design-Systems-Coalition/events/257976364/)
 
----
-
-
-### Shaping design systems
 
 ![](https://user-images.githubusercontent.com/334891/50564566-e16b4280-0cf3-11e9-8804-1a57fb157aa0.jpg)
+
+### Shaping design systems
 
 December 5 @ Datadog
 
 Featuring talks by Derek Howles from Datadog, Nicole Yeo from Artsy, and Carl Nelson.
 [Meetup](https://www.meetup.com/NYC-Design-Systems-Coalition/events/255728499/)
 
----
-
-### October happy hour sponsored by Reaktor
 
 ![](https://user-images.githubusercontent.com/334891/50576129-e1913e00-0dd8-11e9-8cb6-21df4b5803f6.jpg)
+
+### October happy hour sponsored by Reaktor
 
 October 25 @ The Storehouse
 
 A social event for us to chat over drinks and make new connections in the design systems community.
 [Meetup](https://www.meetup.com/NYC-Design-Systems-Coalition/events/254919672/)
 
----
-
-### Putting your design system into practice
 
 ![](https://user-images.githubusercontent.com/334891/47277051-0e7bd600-d58a-11e8-93ab-94857ada089e.jpg)
+
+### Putting your design system into practice
 
 Sept 27 @ MongoDB
 
 Featuring talks by Deep Shah from MongoDB, Marissa Christy from Paperless Post, and Laura Hahn from Priceline.
 [Meetup](https://www.meetup.com/NYC-Design-Systems-Coalition/events/254103852/)
 
----
-
-### July happy hour sponsored by MongoDB
 
 ![](https://user-images.githubusercontent.com/334891/45930307-7c51c500-bf2c-11e8-97ce-0e1c2231c76f.jpg)
+
+### July happy hour sponsored by MongoDB
 
 July 26 @ The Winslow
 
 A social event for us to chat over drinks and make new connections in the design systems community.
 [Meetup](https://www.meetup.com/NYC-Design-Systems-Coalition/events/251864317/)
 
----
-
-### Design system APIs and the developer experience
 
 ![](https://user-images.githubusercontent.com/334891/43049316-62e7f4d0-8dc3-11e8-9a67-588147a71178.jpg)
+
+### Design system APIs and the developer experience
 
 June 20 @ Betterment
 
 Featuring talks by Arielle Sullivan & Connor McNabb from Betterment, Stephanie Manwaring from XO Group, and Jenn Creighton from ClassPass.
 [Meetup](https://www.meetup.com/NYC-Design-Systems-Coalition/events/251191026/)
 
----
-
-### May happy hour
 
 ![](https://user-images.githubusercontent.com/334891/41512111-a336b338-7251-11e8-93b8-7371c8e50a39.JPG)
+
+### May happy hour
 
 May 23 @ The Winslow
 
 A social event for us to chat over drinks and make new connections in the design systems community.
 [Meetup](https://www.meetup.com/NYC-Design-Systems-Coalition/events/249908974/)
 
----
-
-### Bridging the gap
 
 ![](https://user-images.githubusercontent.com/334891/43049909-f3c894d2-8dcd-11e8-9722-298088313767.jpg)
+
+### Bridging the gap
 
 April 19 @ ustwo
 
 Featuring talks by Pedro Sepulveda from ustwo, Mike Perrotti from Meetup, and Fabian Perez from GitHub.
 [Meetup](https://www.meetup.com/NYC-Design-Systems-Coalition/events/247693556/)
 
----
-
-### Tools for collaboration
 
 ![](https://user-images.githubusercontent.com/334891/38999259-66174c58-43bf-11e8-8ee2-4ec1a6fc136c.jpg)
+
+### Tools for collaboration
 
 March 8 @ Meetup
 
 Featuring talks by Noah Levin from Figma, Natalya Shelburne from The New York Times, Brent Jackson from Priceline.com, and Chesley Andrews from Betterment.
 [Meetup](https://www.meetup.com/NYC-Design-Systems-Coalition/events/247534331/)
 
----
-
-### February happy hour
 
 ![](https://user-images.githubusercontent.com/334891/36069622-7d3d42fe-0ebb-11e8-966e-7cd99aa1f925.jpg)
+
+### February happy hour
 
 Feb 7 @ The Winslow
 
@@ -122,62 +117,55 @@ A social event for us to chat over drinks and make new connections in the design
 [Meetup](https://www.meetup.com/NYC-Design-Systems-Coalition/events/247261426/)
 
 
----
-
-### Defining design principles
 
 ![](https://user-images.githubusercontent.com/334891/35812571-627904d6-0a5f-11e8-942a-8e0f7e28af58.jpg)
+
+### Defining design principles
 
 Oct 4 @ The New York Times
 
 Featuring talks by Orr Shtuhl from Wirecutter, Caitlin Osbahr from Buzzfeed, Michael Jovel from the F.D.A., and Magera Holton from Related Works.
 [YouTube](https://youtu.be/VtTbIXnq24k)
 
----
+
+![](/2017-08-DSC-DoSomething-7739.jpg)
 
 ### Empowering product growth with design systems
 
 Aug 17 @ DoSomething
 
 Featuring talks by Una Kravets from Digital Ocean, Luke Patton from DoSomething, and Jina Anne—consultant and community builder.
-
-![](/2017-08-DSC-DoSomething-7739.jpg)
-
 [link](/2017-august)
 
----
+
+![](/may-speaker-photo2.jpg)
 
 ### Documenting design systems
 
 May 11 @ JustWorks
 
 Featuring talks by Tom Takigayama from Justworks, Adekunle Oduye from Nasdaq, and Amélie Lamont from The New York Times.
-
-![](/may-speaker-photo2.jpg)
-
 [link](/2017-may)
 
----
+
+![](/DSC-02Etsy-4469.jpg)
 
 ### Brand design systems that scale
 
 March 23 Etsy @ HQ
 
 Featuring talks by Sophie Shepherd from GitHub, Will Miner from 2U, and Marco Suarez from Etsy.
-
-![](/DSC-02Etsy-4469.jpg)
-
 [link](/2017-march)
 
----
+
+![](/video-thumbnail.jpg)
 
 ### Sharing design systems journeys
 
 December 15 XO @ Group
 
 Featuring Emily Brick from Buzzfeed, Silvia Tueros-Cossio and Philip Kelly from XO Group, and Diana Mounter from GitHub.
-
-![](/video-thumbnail.jpg)
-
 [link](/2016-december)
+
+</Cards>
